fix: incorrect cookie header
I used to just echo back all the cookies I got from the website which used to work fine. But recently the cookies changed and the website sends back cookies to expire right away and when I send those back I'm not authenticated anymore. I can't download any maps when I'm not authenticed so I have to parse the cookie header and only send back the right cookies (the ones that are not expired).